Comment out edge-to-edge display code
The edge-to-edge display feature, which was previously enabled by setting `WindowCompat.setDecorFitsSystemWindows(window, false)` in `MainActivity` and `SplashActivity`, and by setting `android:windowOptOutEdgeToEdgeEnforcement` to true in `styles.xml`, has been commented out.

This change also adjusts the bottom padding of the version text in `SplashActivity` to accommodate the system navigation bar.